requirements.txt
langchain-openai
langchain_community
langchain_core
langchain
streamlit
python-dotenv
langchain-groq
pymupdf
faiss-cpu
langchain_chroma


langchain_huggingface
chromadb
pypdf
langchain-text-splitters

openai
pandas